Minor Issues Fixed

+++ database/migrations/2026_02_19_082025_add_rating_id_to_offers_table.php
@@ -27,8 +27,9 @@ return new class extends Migration {
     public function down(): void
     {
         Schema::table('offers', function (Blueprint $table) {
-            $table->dropForeign(['rating_id']);
-            $table->dropColumn('rating_id');
+            if (Schema::hasColumn('offers', 'rating_id')) {
+                $table->dropColumn('rating_id');
+            }
         });
     }
 };
